requirements.txt
fastapi==0.115.5
uvicorn==0.31.1
redis==5.0.8
httpx==0.27.2
python-dotenv==1.0.1
